fix(client): :bug: fix counter
